pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chain-trace"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "requests>=2.31",
    "python-dateutil>=2.9",
]

[project.scripts]
trace = "chain_trace.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["chain_trace"]
